# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  model: string;
  systemPrompt?: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export type OpenaiModel = (typeof OpenaiModel)[keyof typeof OpenaiModel];

export const OpenaiModel = {
  "gpt-4o-mini": "gpt-4o-mini",
  "gpt-4o": "gpt-4o",
  "o3-mini": "o3-mini",
} as const;

export interface OpenaiModelsResponse {
  models: OpenaiModel[];
}

export interface CreateOpenaiConversationBody {
  title: string;
  model?: OpenaiModel;
  systemPrompt?: string;
}

export interface UpdateOpenaiConversationBody {
  title?: string;
  model?: OpenaiModel;
  systemPrompt?: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  model: string;
  systemPrompt?: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "1536x1024": "1536x1024",
  "1024x1536": "1024x1536",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}