# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns the canonical list of supported model identifiers
 * @summary List supported models
 */
export const ListOpenaiModelsResponse = zod.object({
  models: zod.array(zod.enum(["gpt-4o-mini", "gpt-4o", "o3-mini"])),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  model: zod.string(),
  systemPrompt: zod.string().optional(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
  model: zod.enum(["gpt-4o-mini", "gpt-4o", "o3-mini"]).optional(),
  systemPrompt: zod.string().optional(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  model: zod.string(),
  systemPrompt: zod.string().optional(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Update a conversation's title, model, or system prompt
 */
export const UpdateOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateOpenaiConversationBody = zod.object({
  title: zod.string().optional(),
  model: zod.enum(["gpt-4o-mini", "gpt-4o", "o3-mini"]).optional(),
  systemPrompt: zod.string().optional(),
});

export const UpdateOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  model: zod.string(),
  systemPrompt: zod.string().optional(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Generate an image from a text prompt
 */
export const GenerateOpenaiImageBody = zod.object({
  prompt: zod.string(),
  size: zod.enum(["1024x1024", "1536x1024", "1024x1536"]).optional(),
});

export const GenerateOpenaiImageResponse = zod.object({
  b64_json: zod.string(),
});